Add group fungerer
Add group "#23

+++ Kalender/src/controllere/ConnectionForReal.java
@@ -6,7 +6,7 @@ public class ConnectionForReal {
 
 	public static HTTPConnection scon;
 	public static String name;
-	public static int uid;
+	public static long uid;
 
 	public static void setURL(String url) {
 		scon = new HTTPConnection(url);
